fix msl function's argument types

+++ MetalSwiftSample/add.metal
@@ -8,9 +8,9 @@
 #include <metal_stdlib>
 using namespace metal;
 
-kernel void add_arrays(device const float* inA,
-                       device const float* inB,
-                       device float* result,
+kernel void add_arrays(device const int* inA,
+                       device const int* inB,
+                       device int* result,
                        uint index [[thread_position_in_grid]])
 {
     result[index] = inA[index] + inB[index];
